+++ C7Engine/EntryPoints/CityInteractions.cs
@@ -1,6 +1,3 @@
-using System;
-using System.Collections.Generic;
-
 namespace C7Engine
 {
     using C7GameData;
@@ -15,41 +12,11 @@ namespace C7Engine
             Player owner = gameData.players.Find(player => player.guid == playerGuid);
 
 			Tile tileWithNewCity = MapInteractions.GetTileAt(x, y);
-			City newCity = new City(tileWithNewCity, owner, name, OnUnitCompleted, GetNextItemToBeProduced);
+			City newCity = new City(tileWithNewCity, owner, name);
 			newCity.SetItemBeingProduced(gameData.unitPrototypes["Warrior"]);
             gameData.cities.Add(newCity);
 
             tileWithNewCity.cityAtTile = newCity;
         }
-
-		public static void OnUnitCompleted(MapUnit newUnit)
-		{
-				EngineStorage.gameData.mapUnits.Add(newUnit);
-		}
-
-		public static IProducable GetNextItemToBeProduced(City city, IProducable lastProduced)
-		{
-			Dictionary<string, UnitPrototype> unitPrototypes = EngineStorage.gameData.unitPrototypes;
-			if (lastProduced == unitPrototypes["Warrior"]) {
-				if (city.location.NeighborsCoast()) {
-					Random rng = new Random();
-					if (rng.Next(3) == 0) {
-						return unitPrototypes["Galley"];
-					}
-					else {
-						return unitPrototypes["Chariot"];
-					}
-				}
-				else {
-					return unitPrototypes["Chariot"];
-				}
-			}
-			else if (lastProduced == unitPrototypes["Chariot"]) {
-				return unitPrototypes["Settler"];
-			}
-			else  {
-				return unitPrototypes["Warrior"];
-			}
-		}
     }
 }
